requirements.txt
tensorflow>=2.5.0
numpy>=1.19.5
pandas>=1.3.0
pillow>=8.2.0
matplotlib>=3.4.2
scikit-learn>=0.24.2 